fix: Migrate getCurrentSize() return type to long

+++ server/src/main/java/ru/andrey/kvstorage/server/initialization/impl/SegmentInitializationContextImpl.java
@@ -35,7 +35,7 @@ public class SegmentInitializationContextImpl implements SegmentInitializationCo
     }
 
     @Override
-    public int getCurrentSize() {
+    public long getCurrentSize() {
         return currentSize;
     }
